requirements.txt
discord.py>=2.0.0
python-dotenv>=0.20.0
ollama>=0.1.0
pandas>=1.3.0
pyarrow>=6.0.0
beautifulsoup4>=4.10.0
requests>=2.27.0
aiohttp>=3.8.0
tabulate>=0.9.0
groq